# back-end/tel_openapi.yaml
openapi: "3.0.2"
info:
  description: Control Center API description
  title: Control Center
  version: 1.0.0
paths:
...
  /formal-reports:
    get:
      tags:
        - formalReports
      summary: Get all formal reports available
      operationId: get_freports
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/agencyArray"
        '404':
          description: Not found
        '401':
          description: Not Authorized for this operation
        '500':
          description: Internal server error
        '403':
          description: Forbidden to use this endpoint
  /formal-reports/{formalReportId}:
    get:
      tags:
        - formalReportsId
      summary: Get a specific formal report
      operationId: get_freports_byid
      parameters:
        - name: formalReportId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/formalReportId"
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/formalReport"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
        '400':
          description: Bad request
          
    delete:
      tags:
        - formalReportsId
      summary: Delete a formal report
      operationId: delete_freports_byid
      parameters:
        - name: formalReportId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/formalReportId"
      responses:
        '200':
          description: Successful deletion
        '404':
          description: Not found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
    
    put:
      tags:
        - formalReportsId
      summary: Update a formal report
      operationId: put_freports_byid
      parameters:
        - name: formalReportId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/formalReportId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/formalReport"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
          
    patch:
      tags:
        - formalReportsId
      summary: Update a formal report
      operationId: patch_freports_byid
      parameters:
        - name: formalReportId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/formalReportId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/formalReport"
            example:
              injuries: 10,
              casualties: 2,
              unitsDeployed: 15,
              comments: "come comments"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
...
components:
  schemas:
    agencyId:
      type: string
      description: String containing the agency id
    userId:
      type: string
      description: String containing the user id
    incidentId:
      type: string
      description: String containing the incident id
...
    formalReportId:
      type: string
      description: String containing the formal report id
    user:
      type: object
      required:
        - userId
      properties:
        userId:
          $ref: "#/components/schemas/userId"
        username:
          type: string
          description: Defines unique username, admin has always 'admin' as username
        email:
          type: string
          description: Defines unique email
        password:
          type: string
          minimum: 8
          maximum: 31
          pattern: '^(?=.*[A-Za-z])(?=.*\d)(?=.*[@$!%*#?&])[A-Za-z\d@$!%*#?&]{8,}$'
        role:
          type: string
          description: Role in agency
        agencyId:
          type: string
          description: ID to identify the agency of the user if applicable.
        firstName:
          type: string
        lastName:
          type: string
        age:
          type: number
        gender:
          type: string
        isHeadOfAgency:
          type: boolean
          description: Defines users that represent their corresponding agencies (equals false for normal users)
        x:
          type: number
          description: longtitude
        y:
          type: number
          description: latitude
        isAvailable:
          type: boolean
          description: Used only by normal users (not admin,coordinator,head of agency)
    userArray:
      type: array
      items:
        $ref: "#/components/schemas/user"
...
    formalReport:
      type: object
      required:
        - reportId
      properties:
        formalReportId:
          $ref: "#/components/schemas/formalReportId"
        title:
          type: string
          description: Contains the unique report title
        userId:
          $ref: "#/components/schemas/userId"
        incidentId:
          $ref: "#/components/schemas/incidentId"
        agencyId:
          $ref: "#/components/schemas/agencyId"
        date:
          type: string
          description: The start date of the initial incident
        injuries:
          type: number
        casualties:
          type: number
        unitsDeployed:
          type: number
        comments:
          type: string
          description: String containing the body of the report
        isOpen:
          type: boolean
    agency:
      type: object
      required:
        - agencyId
      properties:
        agencyId:
          $ref: "#/components/schemas/agencyId"
        name:
          type: string
          description: Unique name of agency
        userId:
          $ref: "#/components/schemas/userId"
        address:
          type: string
          description: Physical address of agency (not coordinates)
        users:
          $ref: "#/components/schemas/userArray"
    agencyArray:
      type: array
      items:
        $ref: "#/components/schemas/agency"
...